feat: support specify authentication priority (#26)
This PR adds an opts for calling AccessToken:new, to support specifying a parameter auth_method_order to define which order should the SDK follows. When it is specified with adc, the authentication order will be like Google Cloud SDK's Application Default Credential. When it is not specified or with the value legacy, it follows its old behavior.

This fixes https://konghq.atlassian.net/browse/FTI-6799 by not breaking the existing behavior.